refactor(web): 展示子组件切出（F2）——ChatHeader/Composer/DetailDrawer
- components/ChatHeader.tsx：顶栏玻璃横带（返回/状态行/任务开关/更多菜单/goal 面板/
  接力链条）；moreBtnRef 与 idCopied 为私有展示态随 JSX 下沉；children 插槽承载
  详情抽屉（DOM 上抽屉与顶栏同属 glass-bar，层级保持不变）
- components/DetailDrawer.tsx：详情抽屉三 tab + MCP/context/设置结构化面板；
  McpServerInfo/ContextDataLite/SettingsDataLite 类型随迁导出
- components/Composer.tsx：斜杠面板全套（slashIdx/滚动跟随/键盘导航/Enter 补全）、
  输入框自适应高度、pickImages、codex 模型派生（codexCfg/codexModeOf 等）内聚下沉；
  imgPreviewSrc 导出（send 的回显块仍用）
- Chat.tsx 2052 → 1381 行：props 平铺不分组、无 memo；JSX 逐字搬迁
- 修复：F1 import 编辑误截 Approval 接口字段（typecheck 捕获，已补回）
- typecheck + 406 tests + build 全绿
