+++ web/src/pages/Chat.tsx
@@ -1,56 +1,25 @@
 import { useEffect, useMemo, useRef, useState } from 'react'
 import { createSession, fetchClaudeModelNames, fetchCodexHistory, fetchCodexModels, fetchConfig, fetchHistory, fetchLineage, makeSessionInfo, startHandoff, type CodexModelInfo, type HistoryMessage, type HistoryResponse, type LineageResponse, type ServerConfigInfo, type SessionInfo, type TierModelName } from '../lib/api'
 import { SessionSocket, type CliMsg, type ServerEvent, type SessionState } from '../lib/ws'
-import { StatusPill } from '../components/StatusPill'
 import { ApprovalCard } from '../components/ApprovalCard'
+import { ChatHeader } from '../components/ChatHeader'
+import { Composer, imgPreviewSrc } from '../components/Composer'
+import { DetailDrawer, type ContextDataLite, type McpServerInfo, type SettingsDataLite } from '../components/DetailDrawer'
 import { RewindPicker } from '../components/RewindPicker'
 import { Transcript } from '../components/Transcript'
 import { TasksPanel, type TaskFeed } from '../components/TasksPanel'
-import { ClaudeMark } from '../components/ClaudeMark'
 import { ClaudeStar } from '../components/ClaudeStar'
 import { CodexMark } from '../components/CodexMark'
-import { PopupPanel } from '../components/PopupPanel'
-import { ContextRing } from '../components/ContextRing'
-import { buildTranscriptRows, fmtTokens, nextId, rewindPreview, toolResultText, usageSummary, type Block, type ChatMsg } from '../lib/blocks'
-import { copyText, statusLineOf, cliSidechainToHistory } from '../lib/chatText'
+import { buildTranscriptRows, nextId, rewindPreview, toolResultText, usageSummary, type Block, type ChatMsg } from '../lib/blocks'
+import { statusLineOf, cliSidechainToHistory } from '../lib/chatText'
 import { interceptSlash, type SlashAction } from '../lib/slashIntercept'
 import { appendHistoryMsg, createIngestState, flushStrayResults, hitsSeen, indexToolBlocks, liveMessageKeys, pairToolResultIn, rememberKeys, transcriptKeys, type IngestState, type PendingResult, type ToolPos } from '../lib/ingest'
 import { isCodexKey, isExistingKey } from '../lib/key'
-import { COMMAND_DESC, filterSlashHints, mergeSlashCommands, type SlashEntry } from '../lib/slashCommands'
-
-const MORE_ITEM =
-  'flex w-full items-center gap-2 rounded-[10px] px-3 py-2 text-left font-mono text-[12px] text-muted transition-colors hover:bg-surface hover:text-ink'
 
 interface Approval {
   requestId: string
   toolName: string
   input: unknown
-}
-
-/** claude mcp_status 应答里的单个服务器（buildMcpServerStatuses 形状） */
-interface McpServerInfo {
-  name: string
-  /** connected / failed / disabled / pending / needs-auth 等（CLI 的 connection.type 直出） */
-  status: string
-  error?: string
-  config?: { type?: string; command?: string; args?: string[]; url?: string }
-  scope?: string
-  tools?: { name: string }[]
-}
-
-/** claude get_context_usage 应答的取用子集（analyzeContext 的 ContextData 里我们渲染的部分） */
-interface ContextDataLite {
-  categories: { name: string; tokens: number; isDeferred?: boolean }[]
-  totalTokens: number
-  maxTokens: number
-  percentage: number
-  model?: string
-}
-
-/** claude get_settings 应答的取用子集（settings 全量不枚举——applied + sources 概览 + 原始 JSON 折叠） */
-interface SettingsDataLite {
-  applied?: { model?: string; effort?: string | null }
-  sources?: { source: string; settings: Record<string, unknown> }[]
 }
 
 /** 流式草稿：一轮 assistant 输出的增量块（按 message.id + block index 归并） */
@@ -107,7 +76,6 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
   const [pendingImages, setPendingImages] = useState<
     Array<{ name: string; mediaType: string; dataBase64: string }>
   >([])
-  const fileInputRef = useRef<HTMLInputElement>(null)
   const [detailOpen, setDetailOpen] = useState(false)
   const [detailTitle, setDetailTitle] = useState('')
   const [detailContent, setDetailContent] = useState('加载中…')
@@ -124,12 +92,9 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
   const [goalOpen, setGoalOpen] = useState(false)
   const [goalDraft, setGoalDraft] = useState('')
   const [moreOpen, setMoreOpen] = useState(false)
-  const moreBtnRef = useRef<HTMLButtonElement>(null)
-  const [idCopied, setIdCopied] = useState(false)
   const querySeq = useRef(0)
   const sockRef = useRef<SessionSocket | undefined>(undefined)
   const scrollRef = useRef<HTMLDivElement>(null)
-  const inputRef = useRef<HTMLTextAreaElement>(null)
   const [atBottom, setAtBottom] = useState(true)
   const atBottomRef = useRef(true)
 
@@ -319,37 +284,6 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
       )
       .catch(() => {})
   }, [session.key])
-
-  /** claude 权限模式名 → codex 预设档位（显示用） */
-  const codexModeOf = (m?: string): string => {
-    switch (m) {
-      case 'bypassPermissions':
-        return 'fullAccess'
-      case 'acceptEdits':
-      case 'auto':
-        return 'workspaceAuto'
-      case 'plan':
-        return 'readOnly'
-      case 'readOnly':
-      case 'workspace':
-      case 'workspaceAuto':
-      case 'fullAccess':
-        return m
-      default:
-        return 'workspace'
-    }
-  }
-
-  const codexDefaultModel = codexModels?.find((m) => m.isDefault) ?? codexModels?.[0]
-  const codexModelId = state.model ?? codexDefaultModel?.id
-  const codexCurrentModel = codexModels?.find((m) => m.id === codexModelId) ?? codexDefaultModel
-  const codexEffortLevels: readonly string[] = codexCurrentModel?.efforts.map((e) => e.value) ?? ['low', 'high', 'max']
-  const codexCfg: ServerConfigInfo = {
-    permissionPolicy: 'ask',
-    permissionModes: ['readOnly', 'workspace', 'workspaceAuto', 'fullAccess'],
-    effortLevels: [...codexEffortLevels],
-    models: (codexModels ?? []).map((m) => m.id),
-  }
 
   /** 历史响应落到消息列表 + 从读取位置续订 tail（初次加载与 tail_reset 重载共用） */
   const applyHistory = (resp: HistoryResponse) => {
@@ -1091,17 +1025,6 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
     scheduleFollow(!draft) // 流式进行中走 auto，收尾/新消息才用 smooth
   }, [messages, approvals, draft])
 
-  // 输入框自适应高度：随内容增长，超过 200px 后不再扩大、内部滚动。
-  // 注意 border-box：style.height 包含边框，需补回上下边框宽，否则单行时内容被裁出滚动条
-  useEffect(() => {
-    const el = inputRef.current
-    if (!el) return
-    el.style.height = 'auto'
-    const h = el.scrollHeight + (el.offsetHeight - el.clientHeight)
-    el.style.height = `${Math.min(h, 200)}px`
-    el.style.overflowY = h > 200 ? 'auto' : 'hidden'
-  }, [input])
-
   const onScroll = () => {
     const el = scrollRef.current
     if (!el) return
@@ -1142,28 +1065,6 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
       extra: { serverName, ...(enabled === undefined ? {} : { enabled }) },
     })
   }
-
-  const pickImages = (files: FileList | null) => {
-    if (!files) return
-    for (const f of Array.from(files)) {
-      if (!f.type.startsWith('image/')) continue
-      const reader = new FileReader()
-      reader.onload = () => {
-        const url = String(reader.result ?? '')
-        const comma = url.indexOf(',')
-        if (comma < 0) return
-        setPendingImages((prev) => [
-          ...prev,
-          { name: f.name, mediaType: f.type, dataBase64: url.slice(comma + 1) },
-        ])
-      }
-      reader.readAsDataURL(f)
-    }
-  }
-
-  /** 预览 src：dataURL 与传输用 base64 本是一份数据，渲染时派生 */
-  const imgPreviewSrc = (img: { mediaType: string; dataBase64: string }) =>
-    `data:${img.mediaType};base64,${img.dataBase64}`
 
   // ---------- goal 设定/清除：claude 走 /goal 斜杠命令（本地命令，不进模型上下文），codex 走 thread/goal RPC ----------
   const sendGoal = (condition?: string) => {
@@ -1296,31 +1197,6 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
       })
   }, [showRewind, messages])
 
-  // 优先 initialize 握手返回的命令（含描述），其次 init 消息的命令名，最后空清单（合并层用自有命令兜底）
-  const cliEntries: SlashEntry[] = state.slashCommands?.length
-    ? state.slashCommands.map((c) => ({ name: c.name, desc: c.description }))
-    : (initInfo.slashCommands ?? []).map((n) => ({ name: n, desc: COMMAND_DESC[n] }))
-  // anyplane 自有命令置顶（中文描述优先于 CLI 同名命令），其后是 CLI 报告的完整清单
-  const allEntries = mergeSlashCommands(cliEntries)
-  const slashHints = filterSlashHints(input, allEntries)
-  // 键盘导航：↑↓ 移动，Tab/Enter 采纳，Esc 关闭（索引随清单变化钳位）
-  const [slashIdx, setSlashIdx] = useState(0)
-  const slashActive = slashHints.length > 0 ? Math.min(slashIdx, slashHints.length - 1) : 0
-  /** 面板滚动容器：键盘导航时保证高亮行在视口内 */
-  const slashScrollRef = useRef<HTMLDivElement>(null)
-
-  // 高亮行跟随滚动：只滚面板容器（getBoundingClientRect 相对数学），不动页面滚动条
-  useEffect(() => {
-    const c = slashScrollRef.current
-    if (!c || slashHints.length === 0) return
-    const row = c.querySelectorAll('button')[slashActive]
-    if (!row) return
-    const cRect = c.getBoundingClientRect()
-    const rRect = row.getBoundingClientRect()
-    if (rRect.top < cRect.top) c.scrollTop -= cRect.top - rRect.top
-    else if (rRect.bottom > cRect.bottom) c.scrollTop += rRect.bottom - cRect.bottom
-  }, [slashActive, slashHints.length])
-
   // 渲染行在 Chat 层算：Transcript 保持纯展示，重进/重渲时不重复摊平
   const transcriptRows = useMemo(() => buildTranscriptRows(messages, draft), [messages, draft])
 
@@ -1357,433 +1233,68 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
       </div>
 
       {/* 顶栏：悬浮磨砂横带 */}
-      <div className="glass-bar absolute inset-x-0 top-0 z-30">
-        <div className="px-3 py-2.5">
-          <div className="flex items-center gap-2">
-            <button
-              className="grid h-8 w-8 shrink-0 place-items-center rounded-full bg-surface2 text-muted transition-colors hover:text-ink md:hidden"
-              onClick={props.onBack}
-              title="返回列表"
-              aria-label="返回列表"
-            >
-              <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="h-4 w-4" aria-hidden>
-                <path d="M19 12H5M12 19l-7-7 7-7" />
-              </svg>
-            </button>
-            <div className="min-w-0 flex-1">
-              <div className="truncate text-sm font-medium">{session.title ?? session.cwd ?? session.sessionId}</div>
-              <div className="flex items-center gap-2 font-mono text-[10px] tracking-wide text-faint">
-                <span className={connected ? 'text-ok' : 'text-accent'}>{connected ? '●' : '○'}</span>
-                <span className={busy || phase ? 'text-busy' : ''}>{statusLine}</span>
-              </div>
-            </div>
-            {/* 后台任务侧栏开关：有任务活动时出现；运行中带计数徽标与呼吸 */}
-            {tasks.length > 0 && (
-              <button
-                type="button"
-                className={`relative grid h-8 w-8 shrink-0 place-items-center rounded-full transition-colors ${
-                  tasksOpen ? 'bg-surface2 text-ink' : 'bg-surface2 text-muted hover:text-ink'
-                }`}
-                title="后台任务"
-                aria-label="后台任务"
-                aria-expanded={tasksOpen}
-                onClick={() => setTasksOpen((v) => !v)}
-              >
-                <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="h-4 w-4" aria-hidden>
-                  <path d="M6 3v12" />
-                  <circle cx="18" cy="6" r="3" />
-                  <circle cx="6" cy="18" r="3" />
-                  <path d="M18 9a9 9 0 0 1-9 9" />
-                </svg>
-                {tasks.some((s) => s.status === 'running') && (
-                  <span className="absolute top-0.5 right-0.5 size-1.5 animate-pulse rounded-full bg-busy" aria-hidden />
-                )}
-              </button>
-            )}
-            {(isExisting || !isCodex || state.sessionId) && (
-              <>
-                <button
-                  ref={moreBtnRef}
-                  type="button"
-                  className={`relative grid h-8 w-8 shrink-0 place-items-center rounded-full transition-colors ${
-                    moreOpen ? 'bg-surface2 text-ink' : 'bg-surface2 text-muted hover:text-ink'
-                  }`}
-                  title="更多"
-                  aria-label="更多"
-                  aria-haspopup="menu"
-                  aria-expanded={moreOpen}
-                  onClick={() => setMoreOpen((v) => !v)}
-                >
-                  <svg viewBox="0 0 24 24" fill="currentColor" className="h-4 w-4" aria-hidden>
-                    <circle cx="5" cy="12" r="1.8" />
-                    <circle cx="12" cy="12" r="1.8" />
-                    <circle cx="19" cy="12" r="1.8" />
-                  </svg>
-                  {state.goal && (
-                    <span className="absolute top-0.5 right-0.5 size-1.5 rounded-full bg-ok" aria-hidden />
-                  )}
-                </button>
-                <PopupPanel
-                  open={moreOpen}
-                  anchor={moreBtnRef.current}
-                  onClose={() => setMoreOpen(false)}
-                  placement="bottom-end"
-                  offset={6}
-                  className="min-w-44"
-                >
-                  {currentSessionId && (
-                    <button
-                      type="button"
-                      role="menuitem"
-                      className={`${MORE_ITEM} ${idCopied ? 'text-ok hover:text-ok' : ''}`}
-                      title={`${isCodex ? 'thread id' : 'session id'}：${currentSessionId}（点击复制完整 ID）`}
-                      onClick={() => {
-                        void copyText(currentSessionId).then((ok) => {
-                          if (!ok) {
-                            setMoreOpen(false)
-                            pushSystem(`⚠ 复制失败，请手动复制：${currentSessionId}`, 'error')
-                            return
-                          }
-                          setIdCopied(true)
-                          setTimeout(() => {
-                            setIdCopied(false)
-                            setMoreOpen(false)
-                          }, 900)
-                        })
-                      }}
-                    >
-                      {idCopied ? '✓ 已复制' : `⧉ ${currentSessionId.slice(0, 8)}…`}
-                    </button>
-                  )}
-                  {isExisting && (
-                    <button
-                      type="button"
-                      role="menuitem"
-                      className={MORE_ITEM}
-                      title="会话详情：context 用量 / MCP 状态 / 设置"
-                      onClick={() => {
-                        setMoreOpen(false)
-                        setDetailOpen((v) => !v)
-                        // codex 无 get_context_usage 对应物，默认落在 MCP 状态上
-                        if (!detailOpen) runQuery(isCodex ? 'mcp_status' : 'get_context_usage', isCodex ? 'MCP 状态' : 'context 用量')
-                      }}
-                    >
-                      ▤ 详情
-                    </button>
-                  )}
-                  {(!isCodex || state.sessionId) && (
-                    <button
-                      type="button"
-                      role="menuitem"
-                      className={`${MORE_ITEM} ${state.goal ? 'text-ok hover:text-ok' : ''}`}
-                      title={
-                        state.goal
-                          ? `当前目标：${state.goal.condition}（点击管理）`
-                          : '设定目标：agent 会持续工作直到条件达成（claude /goal · codex thread/goal）'
-                      }
-                      onClick={() => {
-                        setMoreOpen(false)
-                        setGoalDraft(state.goal?.condition ?? '')
-                        setGoalOpen((v) => !v)
-                      }}
-                    >
-                      {state.goal
-                        ? `◎ ${state.goal.condition.slice(0, 16)}${state.goal.condition.length > 16 ? '…' : ''}`
-                        : '◎ 目标'}
-                    </button>
-                  )}
-                  {isExisting && !isCodex && (
-                    <button
-                      type="button"
-                      role="menuitem"
-                      className={MORE_ITEM}
-                      title="分叉当前会话：新分支携带全部历史，原会话保持不动"
-                      onClick={() => {
-                        setMoreOpen(false)
-                        sockRef.current?.send({ kind: 'branch' })
-                      }}
-                    >
-                      ⎇ 分叉
-                    </button>
-                  )}
-                  {isExisting && (
-                    <button
-                      type="button"
-                      role="menuitem"
-                      className={`${MORE_ITEM} disabled:opacity-40`}
-                      disabled={handoffBusy}
-                      title="让另一个 agent 接续本目录的工作（源会话 fork 自写简报，目标会话带简报进场）"
-                      onClick={() => {
-                        setMoreOpen(false)
-                        const toBackend = isCodex ? 'claude' : 'codex'
-                        setHandoffBusy(true)
-                        startHandoff(session.key, toBackend)
-                          .catch((e) => pushSystem(`⚠ 接力失败: ${e instanceof Error ? e.message : e}`, 'error'))
-                          .finally(() => setHandoffBusy(false))
-                      }}
-                    >
-                      {handoffBusy ? '接力中…' : `⇄ 接力给${isCodex ? ' Claude' : ' Codex'}`}
-                    </button>
-                  )}
-                </PopupPanel>
-              </>
-            )}
-          </div>
-          {usageLine && (
-            <div className="mt-1 font-mono text-[10px] tracking-wide text-faint/80">{usageLine}</div>
-          )}
-          {goalOpen && (
-            <div className="mt-2 rounded-[14px] bg-surface2/80 p-2.5 backdrop-blur-xl">
-              <div className="mb-1.5 flex items-center justify-between">
-                <span className="font-mono text-[10px] tracking-wide text-faint">
-                  ◎ 会话目标{state.goal ? '（进行中）' : ''}——agent 会持续工作直到条件达成
-                </span>
-                <button className="font-mono text-[10px] text-faint hover:text-muted" onClick={() => setGoalOpen(false)}>
-                  ✕
-                </button>
-              </div>
-              {state.goal && (
-                <div className="mb-1.5 font-mono text-[11px] leading-relaxed text-ok">
-                  当前：{state.goal.condition}
-                  {state.goal.tokensUsed != null && (
-                    <span className="text-faint"> · {state.goal.tokensUsed} tok</span>
-                  )}
-                </div>
-              )}
-              <div className="flex gap-1.5">
-                <input
-                  className="min-w-0 flex-1 rounded-full bg-bg/60 px-3 py-1.5 font-mono text-[11px] text-ink outline-none placeholder:text-faint/60"
-                  placeholder={isCodex ? '如：迁移完所有调用点并通过测试' : '如：test/auth 全部通过且 lint 干净'}
-                  value={goalDraft}
-                  onChange={(e) => setGoalDraft(e.target.value)}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter' && goalDraft.trim()) {
-                      sendGoal(goalDraft.trim())
-                      setGoalOpen(false)
-                    }
-                  }}
-                />
-                <button
-                  className="shrink-0 rounded-full bg-ink px-3 py-1.5 font-mono text-[11px] text-bg disabled:opacity-40"
-                  disabled={!goalDraft.trim()}
-                  onClick={() => {
-                    if (!goalDraft.trim()) return
-                    sendGoal(goalDraft.trim())
-                    setGoalOpen(false)
-                  }}
-                >
-                  设定
-                </button>
-                {state.goal && (
-                  <button
-                    className="shrink-0 rounded-full px-3 py-1.5 font-mono text-[11px] text-accent hover:bg-accent/10"
-                    onClick={() => {
-                      sendGoal()
-                      setGoalOpen(false)
-                    }}
-                  >
-                    清除
-                  </button>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
-
-        {/* 接力链导航条：仅在当前会话参与血缘时出现 */}
-        {lineage && (
-          <div className="flex items-center gap-1.5 overflow-x-auto px-3 py-1.5 font-mono text-[10px]">
-            <span className="shrink-0 text-faint">⇄ 接力链:</span>
-            {lineage.records
-              .map((r) => {
-                const fromKey = r.fromResolvedKey ?? r.fromKey
-                const toKey = r.toResolvedKey ?? r.toKey
-                const node = (k: string, backend: 'claude' | 'codex') => {
-                  const info = lineage.nodes[k]
-                  const current = k === session.key
-                  return (
-                    <button
-                      key={k}
-                      disabled={!info}
-                      onClick={() => info && props.onNavigate?.(info)}
-                      className={`flex shrink-0 items-center gap-1 rounded-full px-2 py-0.5 ${
-                        current
-                          ? 'bg-surface2 text-ink'
-                          : 'text-faint hover:text-muted'
-                      }`}
-                      title={k}
-                    >
-                      {backend === 'codex' ? <CodexMark size={10} /> : <ClaudeMark className="h-2.5 w-2.5" />}
-                      {new Date(r.at).toLocaleTimeString('zh-CN', { hour: '2-digit', minute: '2-digit' })}
-                    </button>
-                  )
-                }
-                return (
-                  <span key={r.id} className="flex shrink-0 items-center gap-1.5">
-                    {node(fromKey, r.fromBackend)}
-                    <span className="text-faint/60">→</span>
-                    {node(toKey, r.toBackend)}
-                  </span>
-                )
-              })}
-          </div>
-        )}
-
-        {/* 后台任务 Chips 已并入右侧「后台任务」面板（运行中卡片上有停止按钮） */}
-
+      <ChatHeader
+        session={session}
+        connected={connected}
+        statusLine={statusLine}
+        busy={busy}
+        phase={phase}
+        onBack={props.onBack}
+        tasks={tasks}
+        tasksOpen={tasksOpen}
+        onToggleTasks={() => setTasksOpen((v) => !v)}
+        isExisting={isExisting}
+        isCodex={isCodex}
+        sessionId={state.sessionId}
+        currentSessionId={currentSessionId}
+        goal={state.goal}
+        usageLine={usageLine}
+        moreOpen={moreOpen}
+        setMoreOpen={setMoreOpen}
+        onSystemMessage={pushSystem}
+        onToggleDetail={() => {
+          setDetailOpen((v) => !v)
+          // codex 无 get_context_usage 对应物，默认落在 MCP 状态上
+          if (!detailOpen) runQuery(isCodex ? 'mcp_status' : 'get_context_usage', isCodex ? 'MCP 状态' : 'context 用量')
+        }}
+        goalOpen={goalOpen}
+        onToggleGoal={() => {
+          setGoalDraft(state.goal?.condition ?? '')
+          setGoalOpen((v) => !v)
+        }}
+        onCloseGoal={() => setGoalOpen(false)}
+        goalDraft={goalDraft}
+        onGoalDraftChange={setGoalDraft}
+        onSendGoal={sendGoal}
+        onBranch={() => sockRef.current?.send({ kind: 'branch' })}
+        handoffBusy={handoffBusy}
+        onHandoff={() => {
+          const toBackend = isCodex ? 'claude' : 'codex'
+          setHandoffBusy(true)
+          startHandoff(session.key, toBackend)
+            .catch((e) => pushSystem(`⚠ 接力失败: ${e instanceof Error ? e.message : e}`, 'error'))
+            .finally(() => setHandoffBusy(false))
+        }}
+        lineage={lineage}
+        onNavigate={props.onNavigate}
+      >
         {/* 会话详情抽屉 */}
         {detailOpen && (
-          <div className="px-3 py-2">
-            <div className="mb-1.5 flex items-center gap-2 font-mono text-[11px]">
-              <span className="text-muted">{detailTitle}</span>
-              {/* codex 只有 mcp_status 有对应物（mcpServerStatus/list）；context/设置是 claude 控制请求 */}
-              {(isCodex ? (['mcp_status'] as const) : (['get_context_usage', 'mcp_status', 'get_settings'] as const)).map((q) => (
-                <button
-                  key={q}
-                  className="rounded-full bg-surface px-2.5 py-1 text-[10px] text-faint hover:text-ink"
-                  onClick={() =>
-                    runQuery(q, q === 'get_context_usage' ? 'context 用量' : q === 'mcp_status' ? 'MCP 状态' : '设置')
-                  }
-                >
-                  {q === 'get_context_usage' ? 'context' : q === 'mcp_status' ? 'MCP' : '设置'}
-                </button>
-              ))}
-              <button className="ml-auto text-faint hover:text-muted" onClick={() => setDetailOpen(false)}>
-                ✕
-              </button>
-            </div>
-            {detailTitle === 'MCP 状态' && !isCodex && mcpServers ? (
-              /* claude MCP 管理面板：状态 + 重连/启停（toggle 持久化到 settings，与 TUI 同语义） */
-              <div className="max-h-56 overflow-auto rounded-[14px] bg-surface p-2.5">
-                {mcpServers.length === 0 && (
-                  <div className="py-1 font-mono text-[10px] text-faint">无 MCP 服务器（在 claude 配置里添加后出现）</div>
-                )}
-                {mcpServers.map((srv) => {
-                  const meta =
-                    srv.status === 'connected'
-                      ? { dot: 'bg-ok', label: '已连接' }
-                      : srv.status === 'failed'
-                        ? { dot: 'bg-danger', label: '失败' }
-                        : srv.status === 'disabled'
-                          ? { dot: 'bg-faint', label: '已禁用' }
-                          : { dot: 'bg-wait', label: srv.status }
-                  const configLine = srv.config?.url
-                    ? srv.config.url
-                    : srv.config?.command
-                      ? `${srv.config.command} ${(srv.config.args ?? []).join(' ')}`.trim()
-                      : (srv.config?.type ?? '')
-                  const reconnecting = mcpBusy === `${srv.name}:mcp_reconnect`
-                  const toggling = mcpBusy === `${srv.name}:mcp_toggle`
-                  return (
-                    <div key={srv.name} className="flex items-center gap-2 py-1">
-                      <span className={`h-1.5 w-1.5 shrink-0 rounded-full ${meta.dot}`} />
-                      <div className="min-w-0 flex-1">
-                        <div className="font-mono text-[11px] text-ink">
-                          {srv.name}
-                          <span className="text-faint">
-                            {' '}
-                            {meta.label}
-                            {srv.status === 'connected' && srv.tools ? ` · ${srv.tools.length} 工具` : ''}
-                            {srv.scope ? ` · ${srv.scope}` : ''}
-                          </span>
-                        </div>
-                        {configLine && <div className="truncate font-mono text-[10px] text-faint">{configLine}</div>}
-                        {srv.error && <div className="truncate font-mono text-[10px] text-danger">{srv.error}</div>}
-                      </div>
-                      <button
-                        className="shrink-0 rounded-full bg-surface2 px-2.5 py-1 font-mono text-[10px] text-faint hover:text-ink disabled:opacity-40"
-                        disabled={!!mcpBusy || srv.status === 'disabled'}
-                        title="重新连接（mcp_reconnect）"
-                        onClick={() => mcpAction(srv.name, 'mcp_reconnect')}
-                      >
-                        {reconnecting ? '…' : '重连'}
-                      </button>
-                      <button
-                        className="shrink-0 rounded-full bg-surface2 px-2.5 py-1 font-mono text-[10px] text-faint hover:text-ink disabled:opacity-40"
-                        disabled={!!mcpBusy}
-                        title={srv.status === 'disabled' ? '启用并连接（写入 settings）' : '禁用并断开（写入 settings）'}
-                        onClick={() => mcpAction(srv.name, 'mcp_toggle', srv.status === 'disabled')}
-                      >
-                        {toggling ? '…' : srv.status === 'disabled' ? '启用' : '禁用'}
-                      </button>
-                    </div>
-                  )
-                })}
-              </div>
-            ) : detailTitle === 'context 用量' && !isCodex && contextData ? (
-              /* claude context 结构化：总量条 + 分类占比（deferred 类别淡显） */
-              <div className="max-h-56 overflow-auto rounded-[14px] bg-surface p-2.5">
-                <div className="mb-1.5 flex items-baseline justify-between font-mono text-[11px] text-ink">
-                  <span>
-                    {fmtTokens(contextData.totalTokens)} / {fmtTokens(contextData.maxTokens)} tok ·{' '}
-                    {contextData.percentage.toFixed(1)}%
-                  </span>
-                  {contextData.model && (
-                    <span className="text-[10px] text-faint">
-                      {modelNames?.[contextData.model]?.name ?? contextData.model}
-                    </span>
-                  )}
-                </div>
-                <div className="mb-2 h-1 overflow-hidden rounded-full bg-surface2">
-                  <div
-                    className="h-full bg-ink/60"
-                    style={{ width: `${Math.min(100, contextData.percentage)}%` }}
-                  />
-                </div>
-                {contextData.categories.map((c) => (
-                  <div key={c.name} className={`flex items-center gap-2 py-0.5 ${c.isDeferred ? 'opacity-50' : ''}`}>
-                    <span className="w-28 shrink-0 truncate font-mono text-[10px] text-muted" title={c.name}>
-                      {c.name}
-                    </span>
-                    <div className="h-1 flex-1 overflow-hidden rounded-full bg-surface2">
-                      <div
-                        className={`h-full ${c.isDeferred ? 'bg-faint' : 'bg-muted'}`}
-                        style={{
-                          width: `${Math.min(100, (c.tokens / Math.max(1, contextData.maxTokens)) * 100)}%`,
-                        }}
-                      />
-                    </div>
-                    <span className="w-12 shrink-0 text-right font-mono text-[10px] text-faint">
-                      {fmtTokens(c.tokens)}
-                    </span>
-                  </div>
-                ))}
-              </div>
-            ) : detailTitle === '设置' && !isCodex && settingsData ? (
-              /* claude 设置轻结构：生效值 + 来源概览；全量设置不枚举，原始 JSON 折叠兜底 */
-              <div className="max-h-56 overflow-auto rounded-[14px] bg-surface p-2.5">
-                {settingsData.applied && (
-                  <div className="mb-1.5 font-mono text-[11px] text-ink">
-                    当前生效：
-                    <span className="text-muted">
-                      {modelNames?.[settingsData.applied.model ?? '']?.name ?? settingsData.applied.model ?? 'default'}
-                    </span>
-                    <span className="text-faint"> · effort {settingsData.applied.effort ?? '默认'}</span>
-                  </div>
-                )}
-                {(settingsData.sources ?? []).map((s) => (
-                  <div key={s.source} className="flex items-center gap-2 py-0.5 font-mono text-[10px]">
-                    <span className="text-muted">{s.source}</span>
-                    <span className="text-faint">{Object.keys(s.settings ?? {}).length} 项</span>
-                  </div>
-                ))}
-                <details className="mt-1.5">
-                  <summary className="cursor-pointer font-mono text-[10px] text-faint hover:text-muted">
-                    原始 JSON
-                  </summary>
-                  <pre className="mt-1 max-h-40 overflow-auto rounded-[10px] bg-bg/60 p-2 font-mono text-[10px] whitespace-pre-wrap text-muted">
-                    {detailContent}
-                  </pre>
-                </details>
-              </div>
-            ) : (
-              <pre className="max-h-56 overflow-auto rounded-[14px] bg-surface p-2.5 font-mono text-[10px] whitespace-pre-wrap text-muted">
-                {detailContent}
-              </pre>
-            )}
-          </div>
+          <DetailDrawer
+            detailTitle={detailTitle}
+            detailContent={detailContent}
+            isCodex={isCodex}
+            mcpServers={mcpServers}
+            mcpBusy={mcpBusy}
+            onMcpAction={mcpAction}
+            contextData={contextData}
+            settingsData={settingsData}
+            modelNames={modelNames}
+            onRunQuery={runQuery}
+            onClose={() => setDetailOpen(false)}
+          />
         )}
-      </div>
+      </ChatHeader>
 
       {showRewind && (
         <RewindPicker
@@ -1807,247 +1318,52 @@ export function Chat(props: { session: SessionInfo; onBack: () => void; onNaviga
       )}
 
       {/* 输入区：悬浮磨砂圆角块；模型胶囊 / 图片 / 发送全收进块内 */}
-      <div className="absolute inset-x-0 bottom-0 z-30 px-3 pb-3 pt-2">
-        <div className="mx-auto max-w-3xl">
-          {slashHints.length > 0 && (
-            <div className="mb-2 rounded-[14px] bg-surface2/85 p-1 shadow-[0_16px_40px_-12px_rgba(0,0,0,0.5)] backdrop-blur-xl">
-              {/* 完整清单可滚动（CLI initialize 握手报告多少就列多少），自有命令置顶；键盘导航时高亮行跟随滚动 */}
-              <div ref={slashScrollRef} className="max-h-60 overflow-y-auto">
-                {slashHints.map((c, i) => (
-                  <button
-                    key={c.name}
-                    className={`flex w-full items-center gap-2 rounded-[10px] px-2.5 py-1.5 text-left ${
-                      i === slashActive ? 'bg-surface' : 'hover:bg-surface'
-                    }`}
-                    onMouseEnter={() => setSlashIdx(i)}
-                    onClick={() => {
-                      setInput(`/${c.name} `)
-                      setSlashIdx(0)
-                      inputRef.current?.focus()
-                    }}
-                  >
-                    <span className="font-mono text-[12px] text-ink">/{c.name}</span>
-                    {c.desc && <span className="truncate text-xs text-faint">{c.desc}</span>}
-                  </button>
-                ))}
-              </div>
-              <div className="px-2.5 py-1 font-mono text-[9px] tracking-wide text-faint">
-                {slashHints.length} 个命令 · ↑↓ 移动 · Tab 补全
-                {input.trim() === '/' && ' · 继续输入可过滤'}
-              </div>
-            </div>
-          )}
-          <div className="relative">
-            {/* ↓ 与输入块同列、贴在正上方；不放进磨砂块内，否则 backdrop 只能糊到父级内部 */}
-            {!atBottom && (
-              <button
-                className="absolute bottom-full right-0 z-40 mb-2 grid h-9 w-9 place-items-center rounded-full bg-surface2/85 text-ink shadow-lg backdrop-blur-xl hover:bg-surface2"
-                onClick={() => scrollToBottom(false)}
-                title="回到底部"
-                aria-label="回到底部"
-              >
-                ↓
-              </button>
-            )}
-            <div className="rounded-[14px] bg-surface2/80 px-3 pb-2 pt-2.5 shadow-[0_16px_40px_-12px_rgba(0,0,0,0.5)] backdrop-blur-xl">
-            {/* busy 时发送方式：插队（steer，下一边界被模型看到）/ 排队（queue，当前轮结束后） */}
-            {busy && (
-              <div className="mb-1.5 flex items-center gap-1.5 font-mono text-[10px]">
-                <span className="text-faint">工作中，发送：</span>
-                {(['steer', 'queue'] as const).map((m) => (
-                  <button
-                    key={m}
-                    className={`rounded-full px-2.5 py-0.5 ${
-                      sendMode === m ? 'bg-surface text-ink' : 'text-faint hover:text-muted'
-                    }`}
-                    onClick={() => setSendMode(m)}
-                  >
-                    {m === 'steer' ? '插队' : '排队'}
-                  </button>
-                ))}
-                <span className="text-faint/70">
-                  {sendMode === 'steer'
-                    ? isCodex
-                      ? '追加进当前轮'
-                      : '打断当前并立即处理'
-                    : '当前轮结束后自动开始'}
-                </span>
-              </div>
-            )}
-            {/* 待发送图片预览 */}
-            {pendingImages.length > 0 && (
-              <div className="mb-2 flex flex-wrap gap-2">
-                {pendingImages.map((img, i) => (
-                  <span key={i} className="relative">
-                    <img src={imgPreviewSrc(img)} alt={img.name} className="h-14 w-14 rounded-[10px] object-cover" />
-                    <button
-                      className="absolute -right-1.5 -top-1.5 grid h-4 w-4 place-items-center rounded-full bg-accent text-[9px] leading-none text-white"
-                      onClick={() => setPendingImages((prev) => prev.filter((_, j) => j !== i))}
-                    >
-                      ✕
-                    </button>
-                  </span>
-                ))}
-              </div>
-            )}
-            <input
-              ref={fileInputRef}
-              type="file"
-              accept="image/jpeg,image/png,image/gif,image/webp"
-              multiple
-              className="hidden"
-              onChange={(e) => {
-                pickImages(e.target.files)
-                e.target.value = ''
-              }}
-            />
-            <textarea
-              ref={inputRef}
-              className="max-h-[200px] min-h-[1.5rem] w-full resize-none overflow-hidden bg-transparent px-1 text-[15px] leading-snug text-ink outline-none placeholder:text-faint"
-              rows={1}
-              placeholder={busy ? '工作中…' : 'ᕕ( ◠ڼ◠ )ᕗ'}
-              value={input}
-              onChange={(e) => {
-                setInput(e.target.value)
-                setSlashIdx(0)
-              }}
-              onKeyDown={(e) => {
-                // 斜杠命令面板打开时的键盘导航
-                if (slashHints.length > 0) {
-                  if (e.key === 'ArrowDown') {
-                    e.preventDefault()
-                    setSlashIdx((i) => Math.min(i + 1, slashHints.length - 1))
-                    return
-                  }
-                  if (e.key === 'ArrowUp') {
-                    e.preventDefault()
-                    setSlashIdx((i) => Math.max(i - 1, 0))
-                    return
-                  }
-                  if (e.key === 'Tab') {
-                    e.preventDefault()
-                    setInput(`/${slashHints[slashActive].name} `)
-                    setSlashIdx(0)
-                    return
-                  }
-                  if (e.key === 'Escape') {
-                    e.preventDefault()
-                    setInput('')
-                    setSlashIdx(0)
-                    return
-                  }
-                }
-                if (e.key === 'Enter' && !e.shiftKey && !e.nativeEvent.isComposing) {
-                  e.preventDefault()
-                  // 输入还是高亮命令的真前缀时先补全不发送；完整命令名（如 /compact）才直接发送
-                  const trimmed = input.trim()
-                  const active = slashHints[slashActive]
-                  if (slashHints.length > 0 && active && `/${active.name}` !== trimmed) {
-                    setInput(`/${active.name} `)
-                    setSlashIdx(0)
-                    return
-                  }
-                  send()
-                }
-              }}
-            />
-            <div className="mt-1 flex min-w-0 items-center gap-1.5">
-              <div className="min-w-0 flex-1">
-              {cfg && !isCodex && (
-                <StatusPill
-                  cfg={cfg}
-                  model={initInfo.model}
-                  permissionMode={permMode}
-                  effort={effort}
-                  modelNames={modelNames}
-                  onPanelOpen={loadModelNames}
-                  onSetModel={(m) => {
-                    setInitInfo((prev) => ({ ...prev, model: m }))
-                    sockRef.current?.send({ kind: 'control', subtype: 'set_model', extra: { model: m } })
-                  }}
-                  onSetMode={handleSetMode}
-                  onSetEffort={handleSetEffort}
-                />
-              )}
-              {isCodex && codexModels && codexModels.length > 0 && (
-                <StatusPill
-                  cfg={codexCfg}
-                  model={codexModelId}
-                  permissionMode={codexModeOf(permMode ?? state.permissionMode)}
-                  effort={effort ?? state.effort ?? codexCurrentModel?.defaultEffort}
-                  effortLevels={codexEffortLevels}
-                  onSetModel={(m) => {
-                    sockRef.current?.send({ kind: 'control', subtype: 'set_model', extra: { model: m } })
-                  }}
-                  onSetMode={handleSetMode}
-                  onSetEffort={handleSetEffort}
-                />
-              )}
-              </div>
-              {/* 上下文窗口占用环：首个 API 应答/首个 turn 前（state.context 缺省）不渲染 */}
-              <ContextRing
-                backend={isCodex ? 'codex' : 'claude'}
-                context={state.context}
-                usage={state.usage}
-                modelLabel={
-                  isCodex
-                    ? (codexCurrentModel?.label ?? codexModelId)
-                    : (modelNames?.[initInfo.model ?? '']?.name ?? initInfo.model)
-                }
-                onOpenFullDetail={
-                  !isCodex && isExisting
-                    ? () => {
-                        setDetailOpen(true)
-                        runQuery('get_context_usage', 'context 用量')
-                      }
-                    : undefined
-                }
-              />
-              <button
-                type="button"
-                className="grid h-8 w-8 shrink-0 place-items-center rounded-full text-muted transition-colors hover:bg-surface hover:text-ink"
-                title="添加图片（jpg/png/gif/webp，≤5MB）"
-                aria-label="添加图片"
-                onClick={() => fileInputRef.current?.click()}
-              >
-                <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="h-4 w-4" aria-hidden="true">
-                  <rect x="3" y="3" width="18" height="18" rx="2" />
-                  <circle cx="8.5" cy="8.5" r="1.5" />
-                  <path d="m21 15-5-5L5 21" />
-                </svg>
-              </button>
-              {busy ? (
-                <button
-                  type="button"
-                  className="grid h-8 w-8 shrink-0 place-items-center rounded-full bg-accent text-white transition-opacity hover:opacity-85"
-                  onClick={() => sockRef.current?.send({ kind: 'control', subtype: 'interrupt' })}
-                  title="中断当前回合"
-                  aria-label="中断当前回合"
-                >
-                  <svg viewBox="0 0 24 24" fill="currentColor" className="h-4 w-4" aria-hidden="true">
-                    <rect x="5" y="5" width="14" height="14" rx="3" />
-                  </svg>
-                </button>
-              ) : (
-                <button
-                  type="button"
-                  className="grid h-8 w-8 shrink-0 place-items-center rounded-full bg-ink text-bg transition-opacity hover:opacity-85 disabled:pointer-events-none disabled:opacity-25"
-                  disabled={(!input.trim() && pendingImages.length === 0) || !connected}
-                  onClick={send}
-                  title="发送"
-                  aria-label="发送"
-                >
-                  <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.4" strokeLinecap="round" strokeLinejoin="round" className="h-4 w-4" aria-hidden="true">
-                    <path d="M12 19V5" />
-                    <path d="m5 12 7-7 7 7" />
-                  </svg>
-                </button>
-              )}
-            </div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <Composer
+        input={input}
+        onInputChange={setInput}
+        busy={busy}
+        connected={connected}
+        sendMode={sendMode}
+        onSendModeChange={setSendMode}
+        isCodex={isCodex}
+        pendingImages={pendingImages}
+        onPendingImagesChange={setPendingImages}
+        onSend={send}
+        onInterrupt={() => sockRef.current?.send({ kind: 'control', subtype: 'interrupt' })}
+        atBottom={atBottom}
+        onScrollToBottom={() => scrollToBottom(false)}
+        slashCommands={state.slashCommands}
+        initSlashCommands={initInfo.slashCommands}
+        cfg={cfg}
+        claudeModel={initInfo.model}
+        permMode={permMode}
+        effort={effort}
+        modelNames={modelNames}
+        onPanelOpen={loadModelNames}
+        onSetClaudeModel={(m) => {
+          setInitInfo((prev) => ({ ...prev, model: m }))
+          sockRef.current?.send({ kind: 'control', subtype: 'set_model', extra: { model: m } })
+        }}
+        onSetMode={handleSetMode}
+        onSetEffort={handleSetEffort}
+        codexModels={codexModels}
+        stateModel={state.model}
+        statePermissionMode={state.permissionMode}
+        stateEffort={state.effort}
+        onSetCodexModel={(m) => {
+          sockRef.current?.send({ kind: 'control', subtype: 'set_model', extra: { model: m } })
+        }}
+        context={state.context}
+        usage={state.usage}
+        onOpenFullDetail={
+          !isCodex && isExisting
+            ? () => {
+                setDetailOpen(true)
+                runQuery('get_context_usage', 'context 用量')
+              }
+            : undefined
+        }
+      />
       </div>
       <TasksPanel
         open={tasksOpen}
